# docs/api-contract.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Architect Canonical API Contract",
    "version": "1.0.0",
    "x-artifact-type": "CANONICAL_API_CONTRACT",
    "x-owner": "architect"
  },
  "paths": {
...
    "/api/v1/profiles/{username}": {
      "get": {
        "operationId": "Profile_get_api_v1_profiles_username",
        "summary": "Profile",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/profiles/{username}/posts": {
      "get": {
        "operationId": "Profile_get_api_v1_profiles_username_posts",
        "summary": "Profile",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/posts": {
      "post": {
        "operationId": "Posts_post_api_v1_posts",
        "summary": "Posts",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/posts/{postId}": {
      "delete": {
        "operationId": "Posts_delete_api_v1_posts_postId",
        "summary": "Posts",
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/posts/{postId}/replies": {
      "post": {
        "operationId": "Replies_post_api_v1_posts_postId_replies",
        "summary": "Replies",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/posts/{postId}/likes": {
      "post": {
        "operationId": "Likes_post_api_v1_posts_postId_likes",
        "summary": "Likes",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "Likes_delete_api_v1_posts_postId_likes",
        "summary": "Likes",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}